Move things around as per requested
