+++ pandas/io/parsers.py
@@ -970,8 +970,9 @@ class TextFileReader(BaseIterator):
             elif engine == 'python-fwf':
                 klass = FixedWidthFieldParser
             else:
-                raise ValueError('Unknown engine: %r (valid are "c", "python",'
-                                 ' or "python-fwf")' % engine)
+                raise ValueError('Unknown engine: {engine} (valid options are'
+                                 ' "c", "python", or' ' "python-fwf")'.format(
+                                     engine=engine))
             self._engine = klass(self.f, **self.options)
 
     def _failover_to_python(self):
